Restructured Project to follow New File Struct
