+++ conf/cfg.py
@@ -1,0 +1,3 @@
+frames_path = "./assets/frames/"
+emotions = {}
+state = 0
